backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-db-logger==0.1.10
djangorestframework==3.12.4
idna==3.2
install==1.3.4
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
PyJWT==2.1.0
